requirements.txt
ffmpeg-python
faster-whisper
googletrans==3.1.0a0
requests
python-dotenv
elevenlabs
yt-dlp